+++ ts/src/utils.ts
@@ -1,9 +1,8 @@
 import * as bs58 from "bs58";
 import { sha256 } from "crypto-hash";
-import { struct } from "superstruct";
 import assert from "assert";
 import { PublicKey, AccountInfo, Connection } from "@solana/web3.js";
-import { idlAddress } from './idl';
+import { idlAddress } from "./idl";
 
 export const TOKEN_PROGRAM_ID = new PublicKey(
   "TokenkegQfeZyiNwAJbNbGKPFXCWuBvf9Ss623VQ5DA"
@@ -16,9 +15,7 @@ async function getMultipleAccounts(
   Array<null | { publicKey: PublicKey; account: AccountInfo<Buffer> }>
 > {
   const args = [publicKeys.map((k) => k.toBase58()), { commitment: "recent" }];
-  // @ts-ignore
-  const unsafeRes = await connection._rpcRequest("getMultipleAccounts", args);
-  const res = GetMultipleAccountsAndContextRpcResult(unsafeRes);
+  const res = await connection._rpcRequest("getMultipleAccounts", args);
   if (res.error) {
     throw new Error(
       "failed to get info about accounts " +
@@ -70,44 +67,6 @@ async function getMultipleAccounts(
     };
   });
 }
-
-function jsonRpcResult(resultDescription: any) {
-  const jsonRpcVersion = struct.literal("2.0");
-  return struct.union([
-    struct({
-      jsonrpc: jsonRpcVersion,
-      id: "string",
-      error: "any",
-    }),
-    struct({
-      jsonrpc: jsonRpcVersion,
-      id: "string",
-      error: "null?",
-      result: resultDescription,
-    }),
-  ]);
-}
-
-function jsonRpcResultAndContext(resultDescription: any) {
-  return jsonRpcResult({
-    context: struct({
-      slot: "number",
-    }),
-    value: resultDescription,
-  });
-}
-
-const AccountInfoResult = struct({
-  executable: "boolean",
-  owner: "string",
-  lamports: "number",
-  data: "any",
-  rentEpoch: "number?",
-});
-
-const GetMultipleAccountsAndContextRpcResult = jsonRpcResultAndContext(
-  struct.array([struct.union(["null", AccountInfoResult])])
-);
 
 const utils = {
   bs58,